+++ module-scam.c
@@ -663,7 +663,7 @@ static int32_t scam_client_init(struct s_client *cl)
 	cl->reader->card_status = CARD_INSERTED;
 	cl->reader->last_g = cl->reader->last_s = time((time_t *)0);
 
-	cs_log_dbg(D_CLIENT, "scam: last_s=%ld, last_g=%ld", cl->reader->last_s, cl->reader->last_g);
+	cs_log_dbg(D_CLIENT, "scam: last_s=%lld, last_g=%lld", (long long)cl->reader->last_s, (long long)cl->reader->last_g);
 
 	cl->pfd = cl->udp_fd;
